Rename Definition Window to Dictionary Window; add some additional styling to h3 styling
