+++ frontend/src/components/dictionary/dictionary_window.jsx
